# Compiled stateful expression