… CHANGELOG.md
… tick/stateful/eval_function_node_test.go
… tick/stateful/eval_reference_node.go
… tick/stateful/eval_string_node.go
… tick/stateful/eval_unary_node.go
… tick/stateful/eval_unary_node_test.go
… tick/stateful/evaluation_funcs.go
… tick/stateful/expr.go
… tick/stateful/expr_benchmarks_test.go
… tick/stateful/expr_dynamic_test.go
… tick/stateful/expr_test.go
… tick/stateful/node_evaluator.go
… tick/stateful/types.go
… tick/stateful/types_test.go
… tick/stateful_expr.go
… where.go