@nathanielc 2016-04-25
>1. Didn't tested function return type changes - there is need to? If so, do we have function to do so? or should I need to create new one and stub it in?

No, the return type of a function is constant. In fact you could add a method to the Func interface so this is explicit and you can add functions to the constant case.

> 2. Not supported DurationNode - I saw the stateful expression did handle DurationNode, but I can't figure out where it's used - not in BinaryNode and not as single node (ex: EvalNum(DurationNode))

There are no methods or functions that use them yet to my knowledge but it is planned see #169 

> 3. In StatefulExpression we are calling "node.eval" - why so? in the new one we don't call this methods are all tests are passing, are we missing tests?.

Can you point to a line where node.eval is being called? Not sure what you are talking about here.

>4. Creating expression return error - this is new "behaviour", compiling an expression can return an error, there is test for it: TestStatefulExpression_Eval_NotSupportedNode, examples:
passing invalid node to compile, example: commentnode
passing invalid node in binarynode

Seems like there are two kinds of errors: 

1. Errors that cannot be resolved. Aka passing commentnode or some other node that makes no sense and cannot be evaluated
2. Errors that are data dependent and could be resolved if the correct data types are passed in.

For case number one the error should cause the task to fail and stop executing.
For the second case the error should be logged and the task should continue executing.

> 5.@nathanielc - you requested to separate to packages as ast and etc, I didn't do this in this pull request, because it's too much big PR

That's fine we can do it later or not at all if its too much.

> 6. I can fix #490 pretty easily, do you want to?

Yes, go ahead an fix it. Just add a note to this PR that it is fixed.

Haven't looked at the code much  yet but overall the process and explanations are good.

Thanks!