@yosiat 2016-04-25
1. Are you sure about this? the return type is interface{}, what makes sure that future developers won't return sometime int64 and others float64?
2. Ok, so this just confusing to put in the stack eval code (https://github.com/influxdata/kapacitor/blob/master/tick/stateful_expr.go#L94,L95)
3. After a second look, I got confused with the code.. so what is the point of this code? and when it's used? https://github.com/influxdata/kapacitor/blob/master/tick/eval.go#L93
4. TryCompileStatefulExpression will return an error only if the expression couldn't be evaluated (this is the first kinda of ever your described) ever not like type mismatch error which can fixed at runtime. so I think we are ok & agree about this?
5.  This is not too much, this will be an easy work but that will make this PR even harder to CR.
6. I will fix #490 later today, this is just making UnaryNode a dynamic node if it's contains a ReferenceNode

By the way, do you want me to squash commits or something? there is huge number of commits whom are poorly written and doesn't describe the change very well.