@nathanielc 2016-04-25
> Are you sure about this? the return type is interface{}, what makes sure that future developers won't return sometime int64 and others float64?

You do :) I checked current implementations and non have variable return types. Since we are specializing the evaluation now we can make this part of the contract. There are three raw data types we deal with now `bool`, `int64` and `float64`.  So you could add `CallBool`,  `CallInt` and  `CallFloat` methods to the interface to force it. Then if we add more types we can easily do that later.

> Ok, so this just confusing to put in the stack eval code (https://github.com/influxdata/kapacitor/blob/master/tick/stateful_expr.go#L94,L95)

Yes

> After a second look, I got confused with the code.. so what is the point of this code? and when it's used? https://github.com/influxdata/kapacitor/blob/master/tick/eval.go#L93

That is the recursive stack based eval function. Since not all expressions are guaranteed to be a BinaryNode that is the entrypoint to evaluation. If the expression is an NumberNode for example it will get pushed on the stack and then popped as the final result of the expression: `lambda: 1.0` is a valid expression that always evals to float64(1).

>TryCompileStatefulExpression will return an error only if the expression couldn't be evaluated (this is the first kinda of ever your described) ever not like type mismatch error which can fixed at runtime. so I think we are ok & agree about this?

yes, its fine if the compile returns an error for the first case. Can you give an example of how it can fix a type issue at runtime? Seems like there are cases when it is not possible to fix?

>This is not too much, this will be an easy work but that will make this PR even harder to CR.
I will fix #490 later today, this is just making UnaryNode a dynamic node if it's contains a ReferenceNode

In that case let's focus on the PR at hand and get a fix in for this later. It will be beneficial to have a simple PR that fixes a small issue in the new code base so others can begin to see how it works.