@yosiat 2016-04-25
- This sounds really great! let's do this is another PR after this get merged :)
- About ``lambda: 1.0`` it's important to tell that this PR supports both BinaryNode and Single node expression such as ``lambda: 1.0``, if so, what is the need for eval.go? You can see the implementation here:
 - Specializer for single nodes: https://github.com/yosiat/kapacitor/blob/compiled-stateful-expression/tick/eval/node_value_accessor_stateful_expression.go 
 - Here we choose between BinaryNode and single node: https://github.com/yosiat/kapacitor/blob/compiled-stateful-expression/tick/eval/specialized_stateful_expr.go#L65,L76
- An example for fixing issue at runtime, won't be type issue (I was mistaken), it can field which doesn't exists in the scope in the given point but in other points exists.
- Ok

Just to make sure - **This PR is finished and waiting for CR feedback**