@nathanielc 2016-04-25
> Just to make sure - This PR is finished and waiting for CR feedback

I have your branch checkout locally and I am looking through it....