@nathanielc 2016-04-25
> Where this eval is called?

Here is the main entry point:  https://github.com/influxdata/kapacitor/blob/master/tick/stateful_expr.go#L36 

and then its called recursively for example here https://github.com/influxdata/kapacitor/blob/master/tick/stateful_expr.go#L107