@nathanielc 2016-04-25
> Currently, I didn't replaced the stateful expression with the new one.

Is there any case where this new code is used? I didn't see any...