@yosiat 2016-04-28
@nathanielc I finished a lot of refactors and changes! **I am ready for full CR**
I have two more tasks, that will some stuff in the getConstatNodeType and the tests, I don't this is stopper for cr.

- [ ] clean getConstantNodeType, since we have GetType on NodeEvaluator
- [ ] Separating test cases, the test case for stateful expression is too much long and hard to maintain, I want to separate it to "eval_{}_node_test.go" so it will be much clearer.

### Current status
**Code coverage:** 78.2%

**Benchmarks:**
Highlights:

1. We are now allocation free evaluation for all nodes
2. Performance of unary greatly improved
3. Performance of two level deep binary (example: "value" > 0 && "value" < 10)
4. _We don't have benchmarks for it_ but function node evaluation should be faster, because we are caching the creation of arguments node evaluator.

```
name                                                                       old time/op    new time/op    delta
_EvalBool_OneOperator_UnaryNode_BoolNode-4                                   68.1ns ± 1%    16.7ns ± 5%   -75.51%          (p=0.008 n=5+5)
_EvalBool_OneOperator_UnaryNode_ReferenceNode-4                               105ns ± 3%      51ns ± 4%   -51.59%          (p=0.008 n=5+5)
_EvalBool_OneOperator_NumberFloat64_NumberFloat64-4                          41.4ns ± 2%    43.5ns ± 2%    +5.12%          (p=0.008 n=5+5)
_EvalBool_OneOperator_NumberFloat64_NumberInt64-4                            42.7ns ± 3%    43.4ns ± 3%      ~             (p=0.571 n=5+5)
_EvalBool_OneOperator_NumberInt64_NumberInt64-4                              40.1ns ± 3%    41.4ns ± 3%    +3.09%          (p=0.040 n=5+5)
_EvalBool_OneOperator_UnaryNode-4                                             136ns ± 3%      80ns ± 2%   -41.37%          (p=0.008 n=5+5)
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberFloat64-4                   75.6ns ± 3%    74.4ns ± 3%      ~             (p=0.286 n=5+5)
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberInt64-4                     77.7ns ± 6%    78.0ns ± 2%      ~             (p=1.000 n=5+5)
_EvalBool_OneOperator_ReferenceNodeFloat64_ReferenceNodeFloat64-4             121ns ± 2%     118ns ± 3%    -2.97%          (p=0.032 n=5+5)
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeFloat64_NumberFloat64-4    94.1ns ± 3%    93.6ns ± 2%      ~             (p=0.730 n=5+5)
_EvalBool_OneOperatorValueChanges_ReferenceNodeFloat64_NumberFloat64-4       1.25µs ± 3%    1.28µs ± 7%      ~             (p=0.310 n=5+5)
_EvalBool_OneOperator_ReferenceNodeInt64_ReferenceNodeInt64-4                 118ns ± 4%     116ns ± 3%      ~             (p=0.397 n=5+5)
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeInt64_NumberInt64-4        89.4ns ± 4%    90.5ns ± 3%      ~             (p=0.381 n=5+5)
_EvalBool_TwoLevelDeep-4                                                      344ns ± 3%     183ns ± 2%   -46.98%          (p=0.008 n=5+5)
_EvalBool_OneOperatorValueChanges_ReferenceNodeInt64_NumberInt64-4           1.25µs ± 3%    1.32µs ± 8%      ~             (p=0.151 n=5+5)

name                                                                       old alloc/op   new alloc/op   delta
_EvalBool_OneOperator_UnaryNode_BoolNode-4                                    8.00B ± 0%    0.00B ±NaN%  -100.00%          (p=0.008 n=5+5)
_EvalBool_OneOperator_UnaryNode_ReferenceNode-4                               8.00B ± 0%    0.00B ±NaN%  -100.00%          (p=0.008 n=5+5)
_EvalBool_OneOperator_NumberFloat64_NumberFloat64-4                          0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_NumberFloat64_NumberInt64-4                            0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_NumberInt64_NumberInt64-4                              0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_UnaryNode-4                                             8.00B ± 0%    0.00B ±NaN%  -100.00%          (p=0.008 n=5+5)
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberFloat64-4                   0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberInt64-4                     0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_ReferenceNodeFloat64_ReferenceNodeFloat64-4            0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeFloat64_NumberFloat64-4    0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorValueChanges_ReferenceNodeFloat64_NumberFloat64-4       0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_ReferenceNodeInt64_ReferenceNodeInt64-4                0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeInt64_NumberInt64-4        0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_TwoLevelDeep-4                                                     0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorValueChanges_ReferenceNodeInt64_NumberInt64-4           0.00B ±NaN%    0.00B ±NaN%      ~     (all samples are equal)

name                                                                       old allocs/op  new allocs/op  delta
_EvalBool_OneOperator_UnaryNode_BoolNode-4                                     1.00 ± 0%     0.00 ±NaN%  -100.00%          (p=0.008 n=5+5)
_EvalBool_OneOperator_UnaryNode_ReferenceNode-4                                1.00 ± 0%     0.00 ±NaN%  -100.00%          (p=0.008 n=5+5)
_EvalBool_OneOperator_NumberFloat64_NumberFloat64-4                           0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_NumberFloat64_NumberInt64-4                             0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_NumberInt64_NumberInt64-4                               0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_UnaryNode-4                                              1.00 ± 0%     0.00 ±NaN%  -100.00%          (p=0.008 n=5+5)
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberFloat64-4                    0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_ReferenceNodeFloat64_NumberInt64-4                      0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_ReferenceNodeFloat64_ReferenceNodeFloat64-4             0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeFloat64_NumberFloat64-4     0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorValueChanges_ReferenceNodeFloat64_NumberFloat64-4        0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperator_ReferenceNodeInt64_ReferenceNodeInt64-4                 0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorWith11ScopeItem_ReferenceNodeInt64_NumberInt64-4         0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_TwoLevelDeep-4                                                      0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
_EvalBool_OneOperatorValueChanges_ReferenceNodeInt64_NumberInt64-4            0.00 ±NaN%     0.00 ±NaN%      ~     (all samples are equal)
```
_Compared to the specialized stateful expression benchmarks I posted in the initial post _